+++ faker_biology/mol_biol/enzyme_data.py
@@ -18,119 +18,93 @@ enzymes = {
                 "Lactate dehydrogenase": {},
                 "Malate dehydrogenase": {},
                 "Isocitrate dehydrogenase": {},
-                "HMG-CoA reductase": {}
+                "HMG-CoA reductase": {},
             },
             "Category:EC 1.1.2": {},
             "Category:EC 1.1.3": {
                 "Glucose oxidase": {},
                 "L-gulonolactone oxidase": {},
                 "Thiamine oxidase": {},
-                "Xanthine oxidase": {}
+                "Xanthine oxidase": {},
             },
             "Category:EC 1.1.4": {},
             "Category:EC 1.1.5": {},
-            "Category:EC 1.1.99": {}
+            "Category:EC 1.1.99": {},
         },
         "Category:EC 1.2": {
             "Category:EC 1.2.1": {
                 "Acetaldehyde dehydrogenase": {},
                 "Glyceraldehyde 3-phosphate dehydrogenase": {},
-                "Pyruvate dehydrogenase": {}
+                "Pyruvate dehydrogenase": {},
             },
-            "Category:EC 1.2.4": {
-                "Oxoglutarate dehydrogenase": {}
-            }
+            "Category:EC 1.2.4": {"Oxoglutarate dehydrogenase": {}},
         },
         "Category:EC 1.3": {
-            "Category:EC 1.3.1": {
-                "Biliverdin reductase": {}
-            },
+            "Category:EC 1.3.1": {"Biliverdin reductase": {}},
             "Category:EC 1.3.2": {},
-            "Category:EC 1.3.3": {
-                "Protoporphyrinogen oxidase": {}
-            },
+            "Category:EC 1.3.3": {"Protoporphyrinogen oxidase": {}},
             "Category:EC 1.3.5": {},
             "Category:EC 1.3.7": {},
-            "Category:EC 1.3.99": {}
+            "Category:EC 1.3.99": {},
         },
-        "Category:EC 1.4": {
-            "Category:EC 1.4.3": {
-                "Monoamine oxidase": {}
-            }
-        },
+        "Category:EC 1.4": {"Category:EC 1.4.3": {"Monoamine oxidase": {}}},
         "Category:EC 1.5": {
             "Category:EC 1.5.1": {
                 "Dihydrofolate reductase": {},
-                "Methylenetetrahydrofolate reductase": {}
+                "Methylenetetrahydrofolate reductase": {},
             },
             "Category:EC 1.5.3": {
                 "Sarcosine oxidase": {},
-                "Dihydrobenzophenanthridine oxidase": {}
+                "Dihydrobenzophenanthridine oxidase": {},
             },
             "Category:EC 1.5.4": {},
             "Category:EC 1.5.5": {},
             "Category:EC 1.5.7": {},
             "Category:EC 1.5.8": {},
-            "Category:EC 1.5.99": {}
+            "Category:EC 1.5.99": {},
         },
         "Category:EC 1.6": {
             "Category:EC 1.6.1": {},
             "Category:EC 1.6.2": {},
             "Category:EC 1.6.3": {},
             "Category:EC 1.8.1": {},
-            "Category:EC 1.6.5": {
-                "NADH dehydrogenase": {}
-            },
+            "Category:EC 1.6.5": {"NADH dehydrogenase": {}},
             "Category:EC 1.6.6": {},
             "Category:EC 1.18.1": {},
             "Category:EC 1.5.1": {},
-            "Category:EC 1.6.99": {}
+            "Category:EC 1.6.99": {},
         },
         "Category:EC 1.7": {
             "Category:EC 1.7.1": {},
             "Category:EC 1.7.2": {},
-            "Category:EC 1.7.3": {
-                "Urate oxidase": {}
-            },
+            "Category:EC 1.7.3": {"Urate oxidase": {}},
             "Category:EC 1.7.7": {},
-            "Category:EC 1.7.99": {
-                "Nitrite reductase": {},
-                "Nitrate reductase": {}
-            }
+            "Category:EC 1.7.99": {"Nitrite reductase": {}, "Nitrate reductase": {}},
         },
         "Category:EC 1.8": {
             "Category:EC 1.8.1": {
                 "Glutathione reductase": {},
-                "Thioredoxin reductase": {}
+                "Thioredoxin reductase": {},
             },
             "Category:EC 1.8.2": {},
-            "Category:EC 1.8.3": {
-                "Sulfite oxidase": {}
-            },
+            "Category:EC 1.8.3": {"Sulfite oxidase": {}},
             "Category:EC 1.8.4": {},
             "Category:EC 1.8.5": {},
             "EC 2.5.1.18": {},
             "Category:EC 1.8.7": {},
             "Category:EC 1.8.98": {},
-            "Category:EC 1.8.99": {}
+            "Category:EC 1.8.99": {},
         },
         "Category:EC 1.9": {
-            "Category:EC 1.9.3": {
-                "Cytochrome c oxidase": {}
-            },
+            "Category:EC 1.9.3": {"Cytochrome c oxidase": {}},
             "Category:EC 1.9.6": {},
-            "EC 1.9.98.1": {}
+            "EC 1.9.98.1": {},
         },
         "Category:EC 1.10": {
             "Category:EC 1.10.1": {},
-            "Category:EC 1.10.2": {
-                "Coenzyme Q - cytochrome c reductase": {}
-            },
-            "Category:EC 1.10.3": {
-                "Catechol oxidase": {},
-                "Laccase": {}
-            },
-            "Category:EC 1.10.99": {}
+            "Category:EC 1.10.2": {"Coenzyme Q - cytochrome c reductase": {}},
+            "Category:EC 1.10.3": {"Catechol oxidase": {}, "Laccase": {}},
+            "Category:EC 1.10.99": {},
         },
         "Category:EC 1.11": {
             "Category:EC 1.11.1": {
@@ -138,7 +112,7 @@ enzymes = {
                 "Catalase": {},
                 "Myeloperoxidase": {},
                 "Thyroid peroxidase": {},
-                "Glutathione peroxidase": {}
+                "Glutathione peroxidase": {},
             }
         },
         "Category:EC 1.12": {
@@ -147,19 +121,17 @@ enzymes = {
             "Category:EC 1.12.5": {},
             "Category:EC 1.12.7": {},
             "Category:EC 1.12.98": {},
-            "Category:EC 1.12.99": {}
+            "Category:EC 1.12.99": {},
         },
         "Category:EC 1.13": {
-            "Category:EC 1.13.11": {
-                "4-hydroxyphenylpyruvate dioxygenase": {}
-            },
+            "Category:EC 1.13.11": {"4-hydroxyphenylpyruvate dioxygenase": {}},
             "Category:EC 1.13.12": {
                 "Renilla-luciferin 2-monooxygenase": {},
                 "Cypridina-luciferin 2-monooxygenase": {},
                 "Firefly luciferase": {},
                 "Watasenia-luciferin 2-monooxygenase": {},
-                "Oplophorus-luciferin 2-monooxygenase": {}
-            }
+                "Oplophorus-luciferin 2-monooxygenase": {},
+            },
         },
         "Category:EC 1.14": {
             "Cytochrome P450 oxidase": {},
@@ -168,50 +140,30 @@ enzymes = {
                 "CYP2D6": {},
                 "CYP2E1": {},
                 "CYP3A4": {},
-                "Cytochrome P450 oxidase": {}
+                "Cytochrome P450 oxidase": {},
             },
-            "Category:EC 1.14.12": {
-                "Nitric oxide dioxygenase": {}
-            },
-            "Category:EC 1.14.13": {
-                "Nitric oxide synthase": {}
-            },
+            "Category:EC 1.14.12": {"Nitric oxide dioxygenase": {}},
+            "Category:EC 1.14.13": {"Nitric oxide synthase": {}},
             "Category:EC 1.14.14": {
                 "Aromatase": {},
                 "CYP2D6": {},
                 "CYP2E1": {},
-                "CYP3A4": {}
+                "CYP3A4": {},
             },
-            "Category:EC 1.14.16": {
-                "Phenylalanine hydroxylase": {}
-            },
-            "Category:EC 1.14.18": {
-                "Tyrosinase": {}
-            }
+            "Category:EC 1.14.16": {"Phenylalanine hydroxylase": {}},
+            "Category:EC 1.14.18": {"Tyrosinase": {}},
         },
-        "Category:EC 1.15": {
-            "Category:EC 1.15.1": {
-                "Superoxide dismutase": {}
-            }
-        },
-        "Category:EC 1.16": {
-            "Category:EC 1.16.3": {
-                "Ceruloplasmin": {}
-            }
-        },
+        "Category:EC 1.15": {"Category:EC 1.15.1": {"Superoxide dismutase": {}}},
+        "Category:EC 1.16": {"Category:EC 1.16.3": {"Ceruloplasmin": {}}},
         "Category:EC 1.17": {
             "Category:EC 1.17.1": {
                 "Leucoanthocyanidin reductase": {},
                 "Xanthine dehydrogenase": {},
                 "Nicotinate dehydrogenase": {},
-                "4-hydroxy-tetrahydrodipicolinate reductase": {}
+                "4-hydroxy-tetrahydrodipicolinate reductase": {},
             },
-            "Category:EC 1.17.2": {
-                "Nicotinate dehydrogenase (cytochrome)": {}
-            },
-            "Category:EC 1.17.3": {
-                "Xanthine oxidase": {}
-            },
+            "Category:EC 1.17.2": {"Nicotinate dehydrogenase (cytochrome)": {}},
+            "Category:EC 1.17.3": {"Xanthine oxidase": {}},
             "Category:EC 1.17.4": {
                 "Ribonucleotide reductase": {},
                 "Ribonucleoside-triphosphate reductase": {},
@@ -220,53 +172,37 @@ enzymes = {
                 "Vitamin-K-epoxide reductase (warfarin-insensitive)": {},
                 "RRM1": {},
                 "RRM2": {},
-                "RRM2B": {}
+                "RRM2B": {},
             },
-            "Category:EC 1.17.5": {
-                "Caffeine dehydrogenase": {}
-            },
+            "Category:EC 1.17.5": {"Caffeine dehydrogenase": {}},
             "Category:EC 1.17.7": {},
-            "Category:EC 1.17.99": {}
+            "Category:EC 1.17.99": {},
         },
-        "Category:EC 1.18": {
-            "Category:EC 1.18.6": {
-                "Nitrogenase": {}
-            }
-        },
-        "Category:EC 1.19": {
-            "Category:EC 1.19.6": {
-                "Nitrogenase (flavodoxin)": {}
-            }
-        },
+        "Category:EC 1.18": {"Category:EC 1.18.6": {"Nitrogenase": {}}},
+        "Category:EC 1.19": {"Category:EC 1.19.6": {"Nitrogenase (flavodoxin)": {}}},
         "Category:EC 1.20": {
             "Category:EC 1.20.1": {},
             "Category:EC 1.20.2": {},
             "Category:EC 1.20.4": {
                 "Arsenate reductase (glutaredoxin)": {},
-                "Glutaredoxin": {}
+                "Glutaredoxin": {},
             },
             "Category:EC 1.20.9": {},
-            "Category:EC 1.20.99": {}
+            "Category:EC 1.20.99": {},
         },
         "Category:EC 1.21": {
-            "Category:EC 1.21.1": {
-                "Iodotyrosine deiodinase": {}
-            },
+            "Category:EC 1.21.1": {"Iodotyrosine deiodinase": {}},
             "Category:EC 1.21.3": {
                 "Isopenicillin N synthase": {},
-                "Tetrahydrocannabinolic acid synthase": {}
+                "Tetrahydrocannabinolic acid synthase": {},
             },
             "Category:EC 1.21.4": {},
             "Category:EC 1.21.99": {
                 "Thyroxine 5-deiodinase": {},
-                "Iodothyronine deiodinase": {}
-            }
+                "Iodothyronine deiodinase": {},
+            },
         },
-        "Category:EC 1.97": {
-            "Category:EC 1.97.1": {
-                "Deiodinase": {}
-            }
-        }
+        "Category:EC 1.97": {"Category:EC 1.97.1": {"Deiodinase": {}}},
     },
     "Category:Transferases": {
         "Glutathione S-transferase": {},
@@ -274,55 +210,45 @@ enzymes = {
             "Category:EC 2.1.1": {
                 "Catechol-O-methyl transferase": {},
                 "DNA methyltransferase": {},
-                "Histone methyltransferase": {}
+                "Histone methyltransferase": {},
             },
             "Category:EC 2.1.3": {
                 "Aspartate transcarbamoylase": {},
-                "Ornithine transcarbamoylase": {}
-            }
+                "Ornithine transcarbamoylase": {},
+            },
         },
         "Category:EC 2.2": {
             "Category:EC 2.2.1": {
                 "Transketolase": {},
                 "Transaldolase": {},
                 "Acetolactate synthase": {},
-                "2-Succinyl-5-enolpyruvyl-6-hydroxy-3-cyclohexene-1-carboxylic-acid synthase": {}
+                "2-Succinyl-5-enolpyruvyl-6-hydroxy-3-cyclohexene-1-carboxylic-acid synthase": {},
             }
         },
         "Category:EC 2.3": {
             "Category:EC 2.3.1": {
                 "Aminolevulinic acid synthase": {},
-                "Choline acetyltransferase": {}
+                "Choline acetyltransferase": {},
             },
             "Category:EC 2.3.2": {
                 "Factor XIII": {},
                 "Gamma glutamyl transpeptidase": {},
-                "Transglutaminase": {}
-            }
+                "Transglutaminase": {},
+            },
         },
         "Category:EC 2.4": {
-            "Category:EC 2.4.2": {
-                "Hypoxanthine-guanine phosphoribosyltransferase": {}
-            },
+            "Category:EC 2.4.2": {"Hypoxanthine-guanine phosphoribosyltransferase": {}},
             "Category:EC 2.5": {},
-            "Category:EC 2.5.1": {
-                "Thiaminase": {}
-            }
+            "Category:EC 2.5.1": {"Thiaminase": {}},
         },
-        "Category:EC 2.5": {
-            "Flavin prenyltransferase": {}
-        },
+        "Category:EC 2.5": {"Flavin prenyltransferase": {}},
         "Category:EC 2.6": {
             "Category:EC 2.6.1": {
                 "Alanine transaminase": {},
-                "Aspartate transaminase": {}
+                "Aspartate transaminase": {},
             }
         },
-        "Category:EC 2.7": {
-            "Category:EC 2.7.2": {
-                "Butyrate kinase": {}
-            }
-        },
+        "Category:EC 2.7": {"Category:EC 2.7.2": {"Butyrate kinase": {}}},
         "Category:EC 2.8": {
             "Thiosulfate sulfurtransferase": {},
             "3-mercaptopyruvate sulfurtransferase": {},
@@ -338,12 +264,12 @@ enzymes = {
             "Molybdopterin synthase": {},
             "tRNA-uridine 2-sulfurtransferase": {},
             "tRNA-5-taurinomethyluridine 2-sulfurtransferase": {},
-            "tRNA-5-methyluridine(54) 2-sulfurtransferase": {}
+            "tRNA-5-methyluridine(54) 2-sulfurtransferase": {},
         },
         "Category:EC 2.9": {
             "L-seryl-tRNA(Sec) selenium transferase": {},
-            "O-phospho-L-seryl-tRNA(Sec):L-selenocysteinyl-tRNA synthase": {}
-        }
+            "O-phospho-L-seryl-tRNA(Sec):L-selenocysteinyl-tRNA synthase": {},
+        },
     },
     "Category:Hydrolases": {
         "Hydrolytic enzyme": {},
@@ -356,30 +282,21 @@ enzymes = {
                 "Phospholipase": {},
                 "Acetylcholinesterase": {},
                 "Cholinesterase": {},
-                "Lipoprotein lipase": {}
+                "Lipoprotein lipase": {},
             },
-            "Category:EC 3.1.2": {
-                "Ubiquitin carboxy-terminal hydrolase L1": {}
-            },
+            "Category:EC 3.1.2": {"Ubiquitin carboxy-terminal hydrolase L1": {}},
             "Category:EC 3.1.3": {
                 "Phosphatase": {},
                 "Alkaline phosphatase": {},
-                "Fructose bisphosphatase": {}
+                "Fructose bisphosphatase": {},
             },
             "Category:EC 3.1.4": {
                 "Phospholipase": {},
-                "CGMP specific phosphodiesterase type 5": {}
+                "CGMP specific phosphodiesterase type 5": {},
             },
-            "Category:EC 3.1.21": {
-                "Restriction enzyme": {},
-                "Deoxyribonuclease": {}
-            },
-            "Category:EC 3.1.26": {
-                "RNase H": {}
-            },
-            "Category:EC 3.1.27": {
-                "Ribonuclease": {}
-            }
+            "Category:EC 3.1.21": {"Restriction enzyme": {}, "Deoxyribonuclease": {}},
+            "Category:EC 3.1.26": {"RNase H": {}},
+            "Category:EC 3.1.27": {"Ribonuclease": {}},
         },
         "Category:EC 3.2": {
             "Category:EC 3.2.1": {
@@ -390,7 +307,7 @@ enzymes = {
                 "Maltase": {},
                 "Lactase": {},
                 "Beta-galactosidase": {},
-                "Hyaluronidase": {}
+                "Hyaluronidase": {},
             }
         },
         "Category:EC 3.3": {
@@ -405,15 +322,11 @@ enzymes = {
             "Leukotriene-A4 hydrolase": {},
             "Limonene-1,2-epoxide hydrolase": {},
             "Microsomal epoxide hydrolase": {},
-            "Trans-epoxysuccinate hydrolase": {}
+            "Trans-epoxysuccinate hydrolase": {},
         },
         "Category:EC 3.4": {
-            "Category:EC 3.4.11": {
-                "Alanine aminopeptidase": {}
-            },
-            "Category:EC 3.4.15": {
-                "Angiotensin converting enzyme": {}
-            },
+            "Category:EC 3.4.11": {"Alanine aminopeptidase": {}},
+            "Category:EC 3.4.15": {"Angiotensin converting enzyme": {}},
             "Category:EC 3.4.21": {
                 "Serine protease": {},
                 "Chymotrypsin": {},
@@ -430,78 +343,54 @@ enzymes = {
                 "Factor XII": {},
                 "Proteinase K": {},
                 "Tissue plasminogen activator": {},
-                "Protein C": {}
+                "Protein C": {},
             },
-            "Category:EC 3.4.22": {
-                "Separase": {}
-            },
+            "Category:EC 3.4.22": {"Separase": {}},
             "Category:EC 3.4.23": {
                 "Pepsin": {},
                 "Rennet": {},
                 "Renin": {},
                 "Trypsinogen": {},
-                "Plasmepsin": {}
+                "Plasmepsin": {},
             },
-            "Category:EC 3.4.24": {
-                "Matrix metalloproteinase": {}
-            },
-            "Category:EC 3.4.25": {
-                "Metalloendopeptidase": {}
-            }
+            "Category:EC 3.4.24": {"Matrix metalloproteinase": {}},
+            "Category:EC 3.4.25": {"Metalloendopeptidase": {}},
         },
         "Category:EC 3.5": {
-            "Category:EC 3.5.1": {
-                "Urease": {}
-            },
-            "Category:EC 3.5.2": {
-                "Beta-lactamase": {}
-            },
-            "Category:EC 3.5.3": {
-                "Arginase": {}
-            },
+            "Category:EC 3.5.1": {"Urease": {}},
+            "Category:EC 3.5.2": {"Beta-lactamase": {}},
+            "Category:EC 3.5.3": {"Arginase": {}},
             "Category:EC 3.5.4": {
                 "Adenosine deaminase": {},
-                "GTP cyclohydrolase I": {}
+                "GTP cyclohydrolase I": {},
             },
-            "Category:EC 3.5.5": {
-                "Nitrilase": {}
-            }
+            "Category:EC 3.5.5": {"Nitrilase": {}},
         },
         "Category:EC 3.6": {
             "Category:EC 3.6.1": {
                 "Helicase": {},
                 "DnaB helicase": {},
-                "RecQ helicase": {}
+                "RecQ helicase": {},
             },
-            "Category:EC 3.6.3": {
-                "ATPase": {},
-                "NaKATPase": {},
-                "ATP synthase": {}
-            }
+            "Category:EC 3.6.3": {"ATPase": {}, "NaKATPase": {}, "ATP synthase": {}},
         },
-        "Category:EC 3.7": {
-            "Kynureninase": {}
-        },
-        "Category:EC 3.8": {
-            "Haloacetate dehalogenase": {}
-        },
+        "Category:EC 3.7": {"Kynureninase": {}},
+        "Category:EC 3.8": {"Haloacetate dehalogenase": {}},
         "Category:EC 3.9": {
             "Phosphoamidase": {},
             "Protein arginine phosphatase": {},
-            "Phosphohistidine phosphatase": {}
+            "Phosphohistidine phosphatase": {},
         },
         "Category:EC 3.10": {
             "N-sulfoglucosamine sulfohydrolase": {},
-            "Cyclamate sulfohydrolase": {}
+            "Cyclamate sulfohydrolase": {},
         },
         "Category:EC 3.11": {
             "Phosphonoacetaldehyde hydrolase": {},
             "Phosphonoacetate hydrolase": {},
-            "Phosphonopyruvate hydrolase": {}
+            "Phosphonopyruvate hydrolase": {},
         },
-        "Category:EC 3.12": {
-            "Trithionate hydrolase": {}
-        },
+        "Category:EC 3.12": {"Trithionate hydrolase": {}},
         "Category:EC 3.13": {
             "UDP-sulfoquinovose synthase": {},
             "2'-hydroxybiphenyl-2-sulfinate desulfinase": {},
@@ -509,8 +398,8 @@ enzymes = {
             "Carbon disulfide hydrolase": {},
             "(CysO sulfur-carrier protein)-S-L-cysteine hydrolase": {},
             "Carbonyl sulfide hydrolase": {},
-            "S-adenosyl-L-methionine hydrolase (adenosine-forming)": {}
-        }
+            "S-adenosyl-L-methionine hydrolase (adenosine-forming)": {},
+        },
     },
     "Category:Lyases": {
         "Category:EC 4.1": {
@@ -518,42 +407,30 @@ enzymes = {
                 "Ornithine decarboxylase": {},
                 "Uridine monophosphate synthetase": {},
                 "Aromatic-L-amino-acid decarboxylase": {},
-                "RubisCO": {}
+                "RubisCO": {},
             },
-            "Category:EC 4.1.2": {
-                "Fructose-bisphosphate aldolase": {}
-            }
+            "Category:EC 4.1.2": {"Fructose-bisphosphate aldolase": {}},
         },
         "Category:EC 4.2": {
-            "Category:EC 4.2.1": {
-                "Carbonic anhydrase": {},
-                "Tryptophan synthase": {}
-            }
+            "Category:EC 4.2.1": {"Carbonic anhydrase": {}, "Tryptophan synthase": {}}
         },
-        "Category:EC 4.3": {
-            "Category:EC 4.3.1": {
-                "Phenylalanine ammonia-lyase": {}
-            }
-        },
+        "Category:EC 4.3": {"Category:EC 4.3.1": {"Phenylalanine ammonia-lyase": {}}},
         "Category:EC 4.4": {
             "Category:EC 4.4.1": {
                 "Cystathionine gamma-lyase": {},
                 "Cystathionine beta-lyase": {},
-                "Leukotriene C4 synthase": {}
+                "Leukotriene C4 synthase": {},
             }
         },
         "Category:EC 4.5": {
             "Category:EC 4.5.1": {
                 "Dichloromethane dehalogenase": {},
-                "Halohydrin dehalogenase": {}
+                "Halohydrin dehalogenase": {},
             }
         },
         "Category:EC 4.6": {
-            "Category:EC 4.6.1": {
-                "Adenylate cyclase": {},
-                "Guanylate cyclase": {}
-            }
-        }
+            "Category:EC 4.6.1": {"Adenylate cyclase": {}, "Guanylate cyclase": {}}
+        },
     },
     "Category:Isomerases": {
         "Category:EC 5.1": {
@@ -565,7 +442,7 @@ enzymes = {
             "Category:EC 5.1.3": {},
             "UDP-glucose 4-epimerase": {},
             "Category:EC 5.1.99": {},
-            "Methylmalonyl CoA epimerase": {}
+            "Methylmalonyl CoA epimerase": {},
         },
         "Category:EC 5.2": {
             "Category:EC 5.2": {},
@@ -595,34 +472,20 @@ enzymes = {
             "Prolycopene isomerase": {},
             "Retinal isomerase": {},
             "Retinol isomerase": {},
-            "Zeta-carotene isomerase": {}
+            "Zeta-carotene isomerase": {},
         },
         "Category:EC 5.3": {
-            "Category:EC 5.3.3": {
-                "Enoyl CoA isomerase": {}
-            },
-            "Category:EC 5.3.4": {
-                "Protein disulfide isomerase": {}
-            }
+            "Category:EC 5.3.3": {"Enoyl CoA isomerase": {}},
+            "Category:EC 5.3.4": {"Protein disulfide isomerase": {}},
         },
-        "Category:EC 5.4": {
-            "Category:EC 5.4.2": {
-                "Phosphoglucomutase": {}
-            }
-        },
+        "Category:EC 5.4": {"Category:EC 5.4.2": {"Phosphoglucomutase": {}}},
         "Category:EC 5.5": {},
-        "Category:EC 5.99": {
-            "Category:EC 5.99.1": {
-                "Topoisomerase": {}
-            }
-        }
+        "Category:EC 5.99": {"Category:EC 5.99.1": {"Topoisomerase": {}}},
     },
     "Category:Ligases": {
         "Category:EC 6.1": {
             "6-carboxytetrahydropterin synthase": {},
-            "Category:EC 6.1.1": {
-                "FARSB": {}
-            }
+            "Category:EC 6.1.1": {"FARSB": {}},
         },
         "Category:EC 6.2": {
             "Acetate--CoA ligase": {},
@@ -679,20 +542,15 @@ enzymes = {
             "L-firefly luciferin-CoA ligase": {},
             "L-proline-L-prolyl-carrier protein ligase": {},
             "D-alanine-D-alanyl-carrier protein ligase": {},
-            "E1 SAMP-activating enzyme": {}
+            "E1 SAMP-activating enzyme": {},
         },
         "Category:EC 6.3": {
             "Glutamine synthetase": {},
             "Argininosuccinate synthetase": {},
-            "CTP synthase": {}
+            "CTP synthase": {},
         },
-        "Category:EC 6.4": {
-            "Pyruvate carboxylase": {},
-            "Acetyl-CoA carboxylase": {}
-        },
-        "Category:EC 6.5": {
-            "DNA ligase": {}
-        },
-        "Category:EC 6.6": {}
-    }
+        "Category:EC 6.4": {"Pyruvate carboxylase": {}, "Acetyl-CoA carboxylase": {}},
+        "Category:EC 6.5": {"DNA ligase": {}},
+        "Category:EC 6.6": {},
+    },
 }