+++ faker_biology/physiology/organelles.py
@@ -1,0 +1,43 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Sat Feb 26 11:37:00 2022
+
+@author: richard
+"""
+eukaryote_organelles = {
+    "eukaryotes_major": [
+        {"name": "cell membrane", "distribution": "all"},
+        {"name": "cell wall", "distribution": "plants"},
+        {"name": "chloroplast", "distribution": "plants"},
+        {"name": "endoplasmic reticulum", "distribution": "all"},
+        {"name": "flagellum", "distribution": "some"},
+        {"name": "Golgi apparatus", "distribution": "all"},
+        {"name": "mitochondrion", "distribution": "most"},
+        {"name": "nucleus", "distribution": "all"},
+        {"name": "vacuole", "distribution": "plants"},
+    ],
+    "eukaryotes_minor": [
+        {"name": "acrosome", "distribution": "most"},
+        {"name": "autophagosome", "distribution": "all"},
+        {"name": "centriole", "distribution": "animals"},
+        {"name": "cilium", "distribution": "most"},
+        {"name": "cnidocyst", "distribution": "cnidarians"},
+        {"name": "eyespot apparatus", "distribution": "algae"},
+        {"name": "glycosome", "distribution": "protozoa"},
+        {"name": "glyoxysome", "distribution": "plants"},
+        {"name": "hydrogenosome", "distribution": "all"},
+        {"name": "lysosome", "distribution": "animals"},
+        {"name": "melanosome", "distribution": "animals"},
+        {"name": "mitosome", "distribution": "unicellular"},
+        {"name": "myofibril", "distribution": "animals"},
+        {"name": "nucleolus", "distribution": "most"},
+        {"name": "ocelloid", "distribution": "rare"},
+        {"name": "parenthesome", "distribution": "fungi"},
+        {"name": "peroxisome", "distribution": "all"},
+        {"name": "porpsome", "distribution": "all"},
+        {"name": "ribosome", "distribution": "all"},
+        {"name": "stress granule", "distribution": "most"},
+        {"name": "vesicle", "distribution": "most"},
+    ],
+}