+++ CHANGELOG.md
@@ -5,3 +5,4 @@ All significant changes to the project will be recorded here
 ### Added
 
 - faker.enzyme for enzyme names and categories
+- faker.organelle for subcellular organelles